backend_backup_original/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI LMS Backend - Learning Management System"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "alembic>=1.14.0",
    "aiosmtplib>=3.0.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.0.0",
    "fastapi[standard]>=0.115.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.26.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyotp>=2.9.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.12.1",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "redis>=7.4.0",
    "requests>=2.31.0",
    "sendgrid>=6.10.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "slowapi>=0.1.8",
    "twilio>=8.12.0",
    "uvicorn[standard]>=0.42.0",
]
